address some format issues
